requirements.txt
# Core dependencies
playwright>=1.52.0
opencv-python>=4.11.0
pytesseract>=0.3.10
numpy>=2.2.6
Pillow>=11.2.1

# Web and networking
requests>=2.32.3
httpx>=0.28.1
urllib3>=2.4.0

# AI and ML
transformers>=4.52.4
torch>=2.7.0
tokenizers>=0.21.1

# Configuration and data handling
PyYAML>=6.0.2
python-dotenv>=1.1.0
pydantic>=2.11.5

# Testing and development
pytest>=8.3.5
pytest-cov>=6.1.1
black>=25.1.0
flake8>=7.2.0
mypy>=1.16.0
isort>=6.0.1

# Utilities
tqdm>=4.67.1
colorama>=0.4.6
filelock>=3.18.0
platformdirs>=4.3.8 